% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CCmixture.predict}
\alias{CCmixture.predict}
\title{Competing Cause Prevalence-Incidence Mixture Model Predictions}
\usage{
CCmixture.predict(l1_x, l2_x, pi_x, data, time.points, theta.hat)
}
\arguments{
\item{l1_x}{A vector containing the names of covariates used in the \ifelse{html}{\out{\eqn{\lambda}<sub>1</sub>}}{ \eqn{\lambda_1}} (progression rate) parameter (must match column name(s) in the input data)}

\item{l2_x}{A vector containing the names of covariates used in the \eqn{\lambda_2} (clearance rate) parameter (must match column name(s) in the input data)}

\item{pi_x}{A vector containing the names of covariates used in the \eqn{\pi} parameter (probability of prevalent disease) (must match column name(s) in the input data)}

\item{data}{Data set of covariates from which to make the predictions.}

\item{time.points}{Numeric vector of time points used to make cumulative risk predictions}

\item{theta.hat}{Parameter estimates for the model to be used (output from CCmixture.fit)}
}
\description{
Competing Cause Prevalence-Incidence Mixture Model Predictions
}
