% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cervmix.fit}
\alias{cervmix.fit}
\title{Competing Cause Prevalence-Incidence Mixture Models}
\usage{
cervmix.fit(
  l1_x,
  l2_x,
  pi_x,
  data,
  num.runs = 30,
  short.epsilon = 0.1,
  epsilon = 1e-08,
  silent = T
)
}
\arguments{
\item{l1_x}{A vector containing the names of covariates used in the \ifelse{html}{\out{&lambda<sub>1</sub>}}{ \eqn{\lambda_1}} (progression rate) parameter (must match column name(s) in the input data)}

\item{l2_x}{A vector containing the names of covariates used in the \ifelse{html}{\out{&lambda<sub>2</sub>}}{ \eqn{\lambda_2}} (clearance rate) parameter (must match column name(s) in the input data)}

\item{pi_x}{A vector containing the names of covariates used in the \eqn{\pi} parameter (probability of prevalent disease) (must match column name(s) in the input data)}

\item{data}{Data used to fit the model containing columns for each term in l1_x, l2_x and pi_x. The first three columns must be (1)
left interval, (2) right interval and (3) z indicator for prevalent/incident disease.The data must contain observed prevalent and incident cases, however some cases may be unknown (`NA`).}

\item{num.runs}{Number of runs of the 'Short' EM algorithm used to determine initial values for the EM algorithm. Defaults to 30.}

\item{short.epsilon}{Convergence criteria used in the 'Short' EM algorithm to determine initial values. Defaults to 0.1.}

\item{epsilon}{Convergence criteria for the change in log-likelihood value used for stopping the EM algorithm. Defaults to 1e-08.}

\item{silent}{Indicator variable for whether to print out each iteration or not. Default value set to FALSE.}
}
\value{
The output is a list containing the following elements:
\itemize{
\item initial.values - initial values determined by the Short EM algorithm process
\item theta.hat - optimum parameter values estimated by the EM algorithm
\item num.iterations - number of iterations until the EM algorithm converged
\item log.likelihood - value of the log.likelihood at the optimum parameter values
\item hess - hessian matrix
\item std.dev - standard deviation of parameter estimates
\item summary - data frame with estimate, std.dev, and 95\% CI for each parameter (useful for data set comparisons)
}
}
\description{
This function fits Competing Cause Prevalence-Incidence mixture models to interval-censored cervical cancer screening data and obtains parameter estimates.
It is possible for the user to select the covariates that will be used for each parameter.
}
\examples{
sim_dat <- cervmix.simulator(2000, c("hpv"), c(), c(),
                     c(exp(-5), -3, 2, -2, 0.25), show_prob = 0.9, i=5)
cervmix.fit(c("hpv"), c(), c(), sim_dat, silent=FALSE)
}
