% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createIntervalData.R
\name{create.interval.data}
\alias{create.interval.data}
\title{Create interval censored data}
\usage{
create.interval.data(patient_info, hist, cyt, hpv)
}
\arguments{
\item{patient_info}{Data frame containing patient information such as Idwoman, Study, Danasc, Age, Darecl, Arm, DaCens, StCens}

\item{hist}{Data frame containing resutls from histological tests for patients. Must have column names Study, Idwoman, VisitHisto, Idhisto, Datehisto, Typehisto, Istorandom, Classhisto.}

\item{cyt}{Data frame containing results from cytologyical tests for patients. Must have column names Study, Idwoman, VisitCyto, Idtest, HPVpres, cyto_idHPV, Datecyto, Type, CytMed, Classcyto, Racctest, FUP}

\item{hpv}{Data frame containing results from HPV tests for patients}
}
\value{
A data frame containing the left and right intervals of CIN2+ detection for women who were HPV+ at baseline.
}
\description{
Create interval censored data
}
