% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{survmix.simulator}
\alias{survmix.simulator}
\title{Simulates cervical cancer screening data with user-specified parameters. Useful for validating that the model is able to recover the true parameter values. Currently it is only possible to simulate
data with the baseline covariates age (older or younger than 40), HPV genotype (HPV16 positive or negative), and cytology (normal/ abnormal).}
\usage{
survmix.simulator(n, l1_x, l2_x, pi_x, params, show_prob = 0.9, i = 5)
}
\arguments{
\item{n}{Number of women in the simulated data set.}

\item{l1_x}{A vector containing the names of covariates used in the \ifelse{html}{\out{\eqn{\lambda}<sub>1</sub>}}{ \eqn{\lambda_1}} (progression rate) parameter. Options are "age", "HPV16" and "cytology".}

\item{l2_x}{A vector containing the names of covariates used in the \eqn{\lambda_2} (clearance rate) parameter. Options are "age", "HPV16" and "cytology".}

\item{pi_x}{A vector containing the names of covariates used in the \eqn{\pi} parameter (probability of prevalent disease). Options are "age", "HPV16" and "cytology".}

\item{params}{Numerical vector the parameter values to be used in the data simulation (first value is background risk, then l1, l2, pi)}

\item{show_prob}{A value representing the probability of a woman showing up for the screening visit. Defaults to 0.9.}

\item{i}{A value representing the interval between screening rounds (in years). Defaults to 5.}
}
\value{
A data frame containing the left and right interval of CIN2+ detection, the indicator of prevalent disease, age (younger or older than 39, 1=younger),
HPV genotype (HPV16 or other, 1=HPV16), and cytology result (normal or abnormal, 1=abnormal).
}
\description{
Simulates cervical cancer screening data with user-specified parameters. Useful for validating that the model is able to recover the true parameter values. Currently it is only possible to simulate
data with the baseline covariates age (older or younger than 40), HPV genotype (HPV16 positive or negative), and cytology (normal/ abnormal).
}
